Commit before attempting ReLU

+++ HLS/DNN.cpp
@@ -12,29 +12,29 @@
 #include "weights.h"
 #include <hls_stream.h>
           
-double P1[H];
-double y[H];
-double P2[K];
+float P1[H];
+float y[H];
+float P2[K];
 unsigned short img[785];
 
-double sigmoid(double alpha) {
+double sigmoid(float alpha) {
     return 1 / (1 + pow(2.71828, -alpha));
 }
 
-void dnn(hls::stream<unsigned short int> &image,hls::stream<double> &scores)
+void dnn(hls::stream<unsigned short int> &image,hls::stream<float> &scores)
 {
 
-	LOAD_IMG:for(int n=0; n < 785; n++){
+	LOAD_IMG:for(unsigned short int n=0; n < 785; n++){
 
 		image.read(img[n]);
 
 	}
 
-    double temp1 =0;
-    int iteration = 0;
+    float temp1 =0;
+    unsigned short int iteration = 0;
     //computing P1
-    COMPUTE_P1:for (int col = 0; col < H; col++) {
-        COMPUTE_P1_INNER:for (int row = 0; row < DIN + 1; row++) {
+    COMPUTE_P1:for (unsigned short int col = 0; col < H; col++) {
+        COMPUTE_P1_INNER:for (unsigned short int row = 0; row < DIN + 1; row++) {
             temp1 += W1[row][col] * img[row];
         }
         P1[iteration] = temp1;
@@ -42,15 +42,15 @@ void dnn(hls::stream<unsigned short int> &image,hls::stream<double> &scores)
         iteration++;
     }
 
-    COMPUTE_y:for (int col = 0; col < H; col++) {
+    COMPUTE_y:for (unsigned short int col = 0; col < H; col++) {
         y[col] = sigmoid(P1[col]);
     }
 
     //computing P2
     iteration = 0;
     temp1 = 0;
-    COMPUTE_P2:for (int col = 0; col < K; col++) {
-        COMPUTE_P2_INNER:for (int row = 0; row < H; row++) {
+    COMPUTE_P2:for (unsigned short int col = 0; col < K; col++) {
+        COMPUTE_P2_INNER:for (unsigned short int row = 0; row < H; row++) {
             temp1 += W2[row][col] * y[row];
         }
         P2[iteration] = temp1;
@@ -58,7 +58,7 @@ void dnn(hls::stream<unsigned short int> &image,hls::stream<double> &scores)
         iteration++;
     }
 
-    OUTPUT:for (int i = 0; i < 10;i++) {
+    OUTPUT:for (unsigned short int i = 0; i < 10;i++) {
 #ifndef __SYNTHESIS__
         std::cout << P2[i] << std::endl;
 #endif
