+++ HLS/functions.h
@@ -1,2 +1,2 @@
 #include <hls_stream.h>
-void dnn(hls::stream<unsigned short int> &image,hls::stream<double> &scores);
+void dnn(hls::stream<unsigned short int> &image,hls::stream<float> &scores);
